+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # UPCOMING
 * Add host check.
 * Shows an alert after successful collecting book info.
-* Support HyRead and Pubu.
+* Support HyRead, Pubu and Google Play Books.
 
 # v0.1
 * Initial version.